+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 ### Added
 - basis DSL, creating/reading comments for table columns
 - creating/reading comments for tables
+- generating HTML file with table comments
 
 ## [0.1.0] - 2023-08-01
 - Start project